Update Skypilot integration package
