+++ src/zenml/integrations/skypilot_aws/flavors/__init__.py
@@ -11,16 +11,16 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-"""Initialization of the Skypilot Azure ZenML orchestrator."""
+"""Skypilot integration flavor for Skypilot AWS orchestrator."""
 
-from zenml.integrations.skypilot.orchestrators.skypilot_base_vm_orchestrator import (  # noqa
-    SkypilotBaseOrchestrator,
-)
-from zenml.integrations.vm_azure.orchestrators.skypilot_azure_vm_orchestrator import (  # noqa
-    SkypilotAzureOrchestrator,
+from zenml.integrations.skypilot_aws.flavors.skypilot_orchestrator_aws_vm_flavor import (
+    SkypilotAWSOrchestratorConfig,
+    SkypilotAWSOrchestratorFlavor,
+    SkypilotAWSOrchestratorSettings,
 )
 
 __all__ = [
-    "SkypilotBaseOrchestrator",
-    "SkypilotAzureOrchestrator",
+    "SkypilotAWSOrchestratorConfig",
+    "SkypilotAWSOrchestratorFlavor",
+    "SkypilotAWSOrchestratorSettings",
 ]
