+++ src/zenml/integrations/skypilot_aws/flavors/skypilot_orchestrator_aws_vm_flavor.py
@@ -11,7 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-"""Skypilot orchestrator Azure flavor."""
+"""Skypilot orchestrator AWS flavor."""
 
 from typing import TYPE_CHECKING, Optional, Type
 
@@ -19,32 +19,32 @@ from zenml.integrations.skypilot.flavors.skypilot_orchestrator_base_vm_config im
     SkypilotBaseOrchestratorConfig,
     SkypilotBaseOrchestratorSettings,
 )
-from zenml.integrations.vm_azure import SKYPILOT_AZURE_ORCHESTRATOR_FLAVOR
+from zenml.integrations.skypilot_aws import SKYPILOT_AWS_ORCHESTRATOR_FLAVOR
 from zenml.logger import get_logger
 from zenml.models import ServiceConnectorRequirements
 from zenml.orchestrators import BaseOrchestratorConfig, BaseOrchestratorFlavor
 
 if TYPE_CHECKING:
-    from zenml.integrations.vm_azure.orchestrators import (
-        SkypilotAzureOrchestrator,
+    from zenml.integrations.skypilot_aws.orchestrators import (
+        SkypilotAWSOrchestrator,
     )
 
 
 logger = get_logger(__name__)
 
 
-class SkypilotAzureOrchestratorSettings(SkypilotBaseOrchestratorSettings):
-    """Skypilot orchestrator settings for Azure."""
+class SkypilotAWSOrchestratorSettings(SkypilotBaseOrchestratorSettings):
+    """Skypilot orchestrator settings."""
 
 
-class SkypilotAzureOrchestratorConfig(  # type: ignore[misc] # https://github.com/pydantic/pydantic/issues/4173
-    SkypilotBaseOrchestratorConfig, SkypilotAzureOrchestratorSettings
+class SkypilotAWSOrchestratorConfig(  # type: ignore[misc] # https://github.com/pydantic/pydantic/issues/4173
+    SkypilotBaseOrchestratorConfig, SkypilotAWSOrchestratorSettings
 ):
-    """Skypilot orchestrator config for Azure."""
+    """Skypilot orchestrator config."""
 
 
-class SkypilotAzureOrchestratorFlavor(BaseOrchestratorFlavor):
-    """Flavor for the Skypilot orchestrator for Azure."""
+class SkypilotAWSOrchestratorFlavor(BaseOrchestratorFlavor):
+    """Flavor for the Skypilot AWS orchestrator."""
 
     @property
     def name(self) -> str:
@@ -53,7 +53,7 @@ class SkypilotAzureOrchestratorFlavor(BaseOrchestratorFlavor):
         Returns:
             Name of the orchestrator flavor.
         """
-        return SKYPILOT_AZURE_ORCHESTRATOR_FLAVOR
+        return SKYPILOT_AWS_ORCHESTRATOR_FLAVOR
 
     @property
     def service_connector_requirements(
@@ -69,7 +69,7 @@ class SkypilotAzureOrchestratorFlavor(BaseOrchestratorFlavor):
             connector is required for this flavor.
         """
         return ServiceConnectorRequirements(
-            resource_type="azure-generic",
+            resource_type="aws-generic",
         )
 
     @property
@@ -97,7 +97,7 @@ class SkypilotAzureOrchestratorFlavor(BaseOrchestratorFlavor):
         Returns:
             The flavor logo.
         """
-        return "https://public-flavor-logos.s3.eu-central-1.amazonaws.com/orchestrator/azure-skypilot.png"
+        return "https://public-flavor-logos.s3.eu-central-1.amazonaws.com/orchestrator/aws-skypilot.png"
 
     @property
     def config_class(self) -> Type[BaseOrchestratorConfig]:
@@ -106,17 +106,17 @@ class SkypilotAzureOrchestratorFlavor(BaseOrchestratorFlavor):
         Returns:
             The config class.
         """
-        return SkypilotAzureOrchestratorConfig
+        return SkypilotAWSOrchestratorConfig
 
     @property
-    def implementation_class(self) -> Type["SkypilotAzureOrchestrator"]:
+    def implementation_class(self) -> Type["SkypilotAWSOrchestrator"]:
         """Implementation class for this flavor.
 
         Returns:
             Implementation class for this flavor.
         """
-        from zenml.integrations.vm_azure.orchestrators import (
-            SkypilotAzureOrchestrator,
+        from zenml.integrations.skypilot_aws.orchestrators import (
+            SkypilotAWSOrchestrator,
         )
 
-        return SkypilotAzureOrchestrator
+        return SkypilotAWSOrchestrator
