+++ src/zenml/integrations/skypilot_aws/orchestrators/__init__.py
@@ -11,16 +11,15 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-"""Skypilot Azure integration flavor initialization."""
+"""Initialization of the Skypilot AWS ZenML orchestrator."""
 
-from zenml.integrations.vm_azure.flavors.skypilot_orchestrator_azure_vm_flavor import (
-    SkypilotAzureOrchestratorConfig,
-    SkypilotAzureOrchestratorFlavor,
-    SkypilotAzureOrchestratorSettings,
+from zenml.integrations.skypilot.orchestrators.skypilot_base_vm_orchestrator import (  # noqa
+    SkypilotBaseOrchestrator,
 )
-
+from zenml.integrations.skypilot_aws.orchestrators.skypilot_aws_vm_orchestrator import (  # noqa
+    SkypilotAWSOrchestrator,
+)
 __all__ = [
-    "SkypilotAzureOrchestratorConfig",
-    "SkypilotAzureOrchestratorFlavor",
-    "SkypilotAzureOrchestratorSettings",
+    "SkypilotBaseOrchestrator",
+    "SkypilotAWSOrchestrator",
 ]
